Separate model binder logic from data logic
The goal here was to:
1. Avoid referencing nanodbc in the model class (as the model will be used by many files which do not need to know about it).
2. Avoid unnecessary source files bloating the codebase - this keeps it header-only while still making sense.

At the expense of generating a header for the binder, which will only be included by that which is loading the file, this feels
like a sufficient solution to cover both cases.

While we're here, we also just swap the static data to be lazily loaded. This is better for encapsulation and avoids static initialization
issues, also technically reducing application startup times.